Added public API to show the left and right utility buttons programmatically (related to #97).
New methods -showLeftUtilityButtonsAnimated: and -showRightUtilityButtonsAnimated: are counterparts to the existing hideUtilityButtonsAnimated:.

Added a toggle switch in the demo app that, when turned on, will exercise this functionality in response to row selection (via tableView:willSelectCellAtIndexPath:).

Also re-organized the code ViewController.m a little bit (some of the UITableView protocol methods were strewn around).