Make utility buttons available through accessibility actions
Assistive technologies like VoiceOver and Switch Control allow users to
perform "actions" on a UI element. To enable that, we implemented an
accessibility action for each utility button. The name of the action is
the same as the accessibility label of the button.

As UIAccessibilityCustomAction class needed for this is available only
on iOS 8.0, care has been taken so that it works no matter with which
SDK it was compiled and on what iOS version it is run.

The underscored methods are underscored so as to make it clear they are
not implementations of UIAccessibilityAction protocol methods, but
rather internal supporting methods.