Convenience accessibility methods for array category
This adds convenience accessibility methods that add accessibility
labels to the utility buttons created with an image wihtout a
title. This is important because buttons with titles automatically get
their title as accessibility label, but buttons without a title and with
an image nearly always need custom accessibility label, as the default
is the image filename stripped of filename suffix. As a matter of fact,
whenever one wants to add an image button using an image method of this
category, they should always use the accessibility version (except when
they already set the correct accessibility label on the UIImage, then it
is not strictly necessary). So best would be to deprecate the "old"
image methods that do not specify accessibility label.

Still, when other accessibility properties (like hint or value) need to
be customized (as well as non-accessibility properties), direct access
to the button after adding it to the array can be used.